requirements.txt
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.0.8
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.0.3
django-filter==2.3.0
django-model-utils==1.3.1
django-six==1.0.4
djangorestframework==3.10.2
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
iniconfig==1.0.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jsonfield==2.1.1
Markdown==3.2.2
more-itertools==8.4.0
numpy==1.19.1
packaging==20.4
pandas==1.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==6.0.0
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-async==0.6.2
rfc3986==1.4.0
screen==1.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
xlwt==1.3.0
